restructured code and fixed null value bug
