+++ lib/core/model/Article.dart
