+++ lib/core/model/ChannelIcon.dart
