+++ lib/core/model/News.dart
