+++ lib/main.dart
@@ -1,11 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:news_reader/ui/home/Home.dart';
+
+
+import 'ui/views/home/Home.dart';
 
 void main() async {
   runApp(MaterialApp(
     debugShowCheckedModeBanner: false,
     theme: ThemeData(fontFamily: 'Nunito'),
-    home: new Home(),
+    home:  Home(),
   ));
 }
