+++ lib/ui/widgets/ArticleRow.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:news_reader/model/Article.dart';
-import 'package:news_reader/ui/NewsWebViewPage.dart';
+import 'package:news_reader/core/model/Article.dart';
+import 'package:news_reader/ui/views/webview/NewsWebViewPage.dart';
+
 import 'package:timeago/timeago.dart';
 
 class ArticleRow extends StatelessWidget {
